+++ packages/ironbean/src/api.ts
@@ -2,7 +2,4 @@ export {
     IPlugin,
     registerPlugin,
     createComponentContext
-} from "./internals";
-export {
-    TestProvider
-} from "./internalsTesting";
+} from "./core/internals";