+++ packages/ironbean/src/core/base.ts
@@ -40,7 +40,7 @@ export class ApplicationContext {
     }
 }
 
-export function getBaseApplicationContext(): ApplicationContext {
+export function getRootAppContext(): ApplicationContext {
     if (containerStorage.mode === StorageMode.Prototype) {
         throw new Error("You use createBaseApplicationContext(), don't use it in combination with getBaseApplicationContext() in the same environment.")
     }
@@ -49,13 +49,12 @@ export function getBaseApplicationContext(): ApplicationContext {
     return container.getBean(ApplicationContext);
 }
 
-export function createBaseApplicationContext(): ApplicationContext {
+export function createRootAppContext(): ApplicationContext {
     if (containerStorage.mode === StorageMode.Singleton) {
         throw new Error("You use getBaseApplicationContext(), don't use it in combination with createBaseApplicationContext() in the same environment.")
     }
     containerStorage.mode = StorageMode.Prototype;
     const container = containerStorage.createBaseContainer();
-
     return container.getBean(ApplicationContext);
 }