+++ packages/ironbean/test/api.spec.ts
@@ -3,19 +3,20 @@ import {
     autowired,
     component,
     ComponentContext,
-    ComponentType, createBaseApplicationContext,
+    ComponentType,
+    createRootAppContext,
     destroyContext,
-    getBaseApplicationContext
+    getRootAppContext
 } from "../src";
-import {Container} from "../src/container";
-import {containerStorage} from "../src/containerStorage";
+import {Container} from "../src/core/container";
+import {containerStorage} from "../src/core/containerStorage";
 import {createComponentContext, IPlugin, registerPlugin} from "../src/api";
 
 describe("api", () => {
     let applicationContext: ApplicationContext;
 
     beforeEach(() => {
-        applicationContext = getBaseApplicationContext();
+        applicationContext = getRootAppContext();
         expectDependenciesCount(2);
     })
 
@@ -78,7 +79,7 @@ describe("api", () => {
     it("plugins for prototype storage mode", () => {
         containerStorage.dispose();
 
-        applicationContext = createBaseApplicationContext();
+        applicationContext = createRootAppContext();
         @component
         class Plugin implements IPlugin {
             componentContext: ComponentContext;