+++ packages/ironbean/test/storage.spec.ts
@@ -1,5 +1,5 @@
-import {createBaseApplicationContext, destroyContext, getBaseApplicationContext} from "../src";
-import {containerStorage} from "../src/containerStorage";
+import {createRootAppContext, destroyContext, getRootAppContext} from "../src";
+import {containerStorage} from "../src/core/containerStorage";
 
 describe("storage", () => {
     afterEach(() => {
@@ -15,24 +15,24 @@ describe("storage", () => {
     });
 
     it("getBaseApplicationContext", () => {
-        const ctx1 = getBaseApplicationContext();
-        const ctx2 = getBaseApplicationContext();
+        const ctx1 = getRootAppContext();
+        const ctx2 = getRootAppContext();
 
         expect(ctx1).toBe(ctx2);
 
         expect(() => {
-            createBaseApplicationContext();
+            createRootAppContext();
         }).toThrowError("You use getBaseApplicationContext(), don't use it in combination with createBaseApplicationContext() in the same environment.");
     });
 
     it("createBaseApplicationContext", () => {
-        const ctx1 = createBaseApplicationContext();
-        const ctx2 = createBaseApplicationContext();
+        const ctx1 = createRootAppContext();
+        const ctx2 = createRootAppContext();
 
         expect(ctx1).not.toBe(ctx2);
 
         expect(() => {
-            getBaseApplicationContext();
+            getRootAppContext();
         }).toThrowError("You use createBaseApplicationContext(), don't use it in combination with getBaseApplicationContext() in the same environment.");
     });
 });