+++ packages/ironbean/test/test.spec.ts
@@ -7,7 +7,7 @@ import {
     ComponentType,
     DependencyToken,
     destroyContext,
-    getBaseApplicationContext,
+    getRootAppContext,
     IFactory,
     lazy,
     needScope,
@@ -19,10 +19,9 @@ import {
     type,
     inject
 } from "../src";
-import {Container} from "../src/container";
-import {containerStorage} from "../src/containerStorage";
-import {CollectionToken} from "../src/collection";
-import {LazyToken} from "../src/lazy";
+import {Container} from "../src/core/container";
+import {containerStorage} from "../src/core/containerStorage";
+import {LazyToken} from "../src/core/lazy";
 abstract class A {
 
 }
@@ -38,7 +37,7 @@ describe("test", () => {
     let applicationContext: ApplicationContext;
 
     beforeEach(() => {
-        applicationContext = getBaseApplicationContext();
+        applicationContext = getRootAppContext();
         expectDependenciesCount(2);
     })
 
@@ -1401,7 +1400,7 @@ describe("test", () => {
                 @autowired
                 item: Object;
             }
-            let context = getBaseApplicationContext();
+            let context = getRootAppContext();
 
             const i = context.getBean(A);
 
@@ -1411,7 +1410,7 @@ describe("test", () => {
         });
 
         it("test1", () => {
-            let context = getBaseApplicationContext();
+            let context = getRootAppContext();
             const oldB = context.getBean(AComponent).b;
             const oldC = context.getBean(AComponent).c;
 
@@ -1429,7 +1428,7 @@ describe("test", () => {
             expect(context.getBean(C)).not.toBe(oldC);
 
             destroyContext();
-            context = getBaseApplicationContext();
+            context = getRootAppContext();
             expect(context.getBean(AComponent).b).not.toBe(oldB);
             expect(context.getBean(B)).not.toBe(oldB);
             expect(context.getBean(AComponent).c).not.toBe(oldC);