+++ packages/ironbean/CHANGELOG.md
@@ -1,5 +1,10 @@
 # ironbean
 
+## 2.0.0
+rename getBaseApplicationContext() to getRootAppContext()
+rename createBaseApplicationContext() to createRootAppContext()
+rename getBaseTestingContext() to getRootTestingContext()
+
 ## 1.0.21
 fixes for E6 environment